userportal/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
PyJWT==2.9.0
Werkzeug==3.0.1
numpy
pandas
numba
python-dotenv
psycopg2-binary
matplotlib
scipy
flask-admin
WTForms==3.1.2
flask-security
flask-bcrypt